voltage reader appears to work

+++ Planter/src/voltage_senor.cpp
@@ -8,7 +8,7 @@ VoltageSensor::VoltageSensor(int analogPin){
 // Calculator method
 float VoltageSensor::take_reading(){
     int adcVal = analogRead(pin);
-    float vOut = (adcVal / ADC_CONVERSION) * VOLTAGE_REF;   // voltage into ADC
-    float voltage = vOut * VOLTAGE_DIVIDER_RATIO;           // device divides voltage by 5
+    float vOut = ((float)adcVal / ADC_CONVERSION) * VOLTAGE_REF;    // voltage into ADC
+    float voltage = vOut * VOLTAGE_DIVIDER_RATIO;                   // device divides voltage by 5
     return voltage;
 }
